Separate lambda_handler helper functions into separate files
