+++ filenameprocessor/tests/test_router_lambda_function.py
@@ -6,11 +6,11 @@ import csv
 from unittest.mock import patch, MagicMock
 import boto3
 from moto import mock_sqs
-from router_lambda_function import (
+from router_lambda_function import create_ack_file
+from send_to_supplier_queue import send_to_supplier_queue
+from initial_file_validation import (
     identify_supplier,
     initial_file_validation,
-    send_to_supplier_queue,
-    create_ack_file,
     validate_vaccine_type_permissions,
     validate_action_flag_permissions,
 )
@@ -39,6 +39,8 @@ class TestRouterLambdaFunctions(unittest.TestCase):
         self.file_key = "Flu_Vaccinations_v5_YGM41_20240708T12130100.csv"
         self.bucket_name = "test-bucket"
         self.ods_code = "YGM41"
+        self.mock_s3_client = MagicMock()
+        self.mock_sqs_client = MagicMock()
 
     def test_identify_supplier(self):
         """tests supplier is correctly matched"""
@@ -51,74 +53,74 @@ class TestRouterLambdaFunctions(unittest.TestCase):
         valid_csv_content_dict_reader = convert_string_to_dict_reader(Constants.valid_file_content)
 
         with (
-            patch("router_lambda_function.get_supplier_permissions", return_value=["FLU_CREATE", "FLU_UPDATE"]),
-            patch("router_lambda_function.get_csv_content_dict_reader", return_value=valid_csv_content_dict_reader),
+            patch("initial_file_validation.get_supplier_permissions", return_value=["FLU_CREATE", "FLU_UPDATE"]),
+            patch("initial_file_validation.get_csv_content_dict_reader", return_value=valid_csv_content_dict_reader),
         ):
-            self.assertTrue(initial_file_validation(self.file_key, self.bucket_name))
+            self.assertTrue(initial_file_validation(self.file_key, self.bucket_name, self.mock_s3_client))
 
-    @patch("router_lambda_function.validate_content_headers")
+    @patch("initial_file_validation.validate_content_headers")
     def test_invalid_extension(self, mock_validate_csv):
         file_key = "Flu_Vaccinations_v5_YGM41_20240708T12130100.txt"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.validate_content_headers")
+    @patch("initial_file_validation.validate_content_headers")
     def test_invalid_file_structure(self, mock_validate_csv):
         file_key = "Flu_Vaccinations_v5_YGM41.csv"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.validate_content_headers")
+    @patch("initial_file_validation.validate_content_headers")
     def test_invalid_vaccine_type(self, mock_validate_csv):
         file_key = "Invalid_Vaccinations_v5_YGM41_20240708T12130100.csv"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.validate_content_headers")
+    @patch("initial_file_validation.validate_content_headers")
     def test_invalid_version(self, mock_validate_csv):
         file_key = "Flu_Vaccinations_v3_YGM41_20240708T12130100.csv"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.validate_content_headers")
+    @patch("initial_file_validation.validate_content_headers")
     def test_invalid_ods_code(self, mock_validate_csv):
         file_key = "Flu_Vaccinations_v5_INVALID_20240708T12130100.csv"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.validate_content_headers")
+    @patch("initial_file_validation.validate_content_headers")
     def test_invalid_timestamp(self, mock_validate_csv):
         file_key = "Flu_Vaccinations_v5_YGM41_20240708Ta99999999.csv"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.get_csv_content_dict_reader")
-    @patch("router_lambda_function.validate_content_headers")
-    def test_invalid_column_count(self, mock_validate_csv, mock_csv_content_dict_reader):
-        mock_csv_content_dict_reader.return_value = Constants.valid_file_content
-        mock_validate_csv.return_value = (False, True)
+    @patch("initial_file_validation.get_csv_content_dict_reader")
+    @patch("initial_file_validation.validate_content_headers")
+    def test_invalid_column_count(self, mock_validate_content_headers, mock_get_csv_content_dict_reader):
+        mock_get_csv_content_dict_reader.return_value = convert_string_to_dict_reader(Constants.valid_file_content)
+        mock_validate_content_headers.return_value = False
         file_key = "Flu_Vaccinations_v5_YGM41_20240708T12130100.csv"
         bucket_name = "test-bucket"
 
-        valid = initial_file_validation(file_key, bucket_name)
+        valid = initial_file_validation(file_key, bucket_name, self.mock_s3_client)
         self.assertFalse(valid)
 
-    @patch("router_lambda_function.sqs_client")
-    def test_supplier_queue_1(self, mock_sqs_client):
+    # @patch("router_lambda_function.sqs_client")
+    def test_supplier_queue_1(self):
         """tests if supplier queue is called"""
-        mock_send_message = mock_sqs_client.send_message
+        mock_send_message = self.mock_sqs_client.send_message
         supplier = "EMIS"
         message_body = {
             "vaccine_type": "Flu",
@@ -126,16 +128,15 @@ class TestRouterLambdaFunctions(unittest.TestCase):
             "timestamp": "20240708T12130100",
             "filename": "Flu_Vaccinations_v5_YGM41_20240708T12130100.csv",
         }
-        send_to_supplier_queue(supplier, message_body)
+        send_to_supplier_queue(supplier, message_body, self.mock_sqs_client)
         mock_send_message.assert_called_once()
 
-    @patch("router_lambda_function.s3_client")
-    def test_create_ack_file(self, mock_s3_client):
+    def test_create_ack_file(self):
         """tests whether ack file is created"""
         validation_passed = True
         created_at_formatted = "20240725T12510700"
-        create_ack_file("1", self.file_key, validation_passed, True, created_at_formatted)
-        mock_s3_client.upload_fileobj.assert_called_once()
+        create_ack_file("1", self.file_key, validation_passed, True, created_at_formatted, self.mock_s3_client)
+        self.mock_s3_client.upload_fileobj.assert_called_once()
 
     @patch.dict(
         os.environ,
@@ -146,15 +147,14 @@ class TestRouterLambdaFunctions(unittest.TestCase):
             "PROD_ACCOUNT_ID": "3456789109",
         },
     )
-    @patch("router_lambda_function.sqs_client")  # Mock the SQS client
-    def test_send_to_supplier_queue(self, mock_sqs_client):
+    def test_send_to_supplier_queue(self):
         """tests SQS queue is sent a message for valid files"""
         # Define the mock SQS queue URL
         mock_url = "https://sqs.eu-west-2.amazonaws.com/123456789012/imms-batch-internal-dev-EMIS-metadata-queue.fifo"
-        mock_sqs_client.get_queue_url = MagicMock(return_value={"QueueUrl": mock_url})
+        self.mock_sqs_client.get_queue_url = MagicMock(return_value={"QueueUrl": mock_url})
 
         mock_send_message = MagicMock()
-        mock_sqs_client.send_message = mock_send_message
+        self.mock_sqs_client.send_message = mock_send_message
 
         supplier = "EMIS"
         message_body = {
@@ -165,7 +165,7 @@ class TestRouterLambdaFunctions(unittest.TestCase):
         }
 
         # Call the send_to_supplier_queue function
-        send_to_supplier_queue(supplier, message_body)
+        send_to_supplier_queue(supplier, message_body, self.mock_sqs_client)
 
         # Assert that send_message was called once
         mock_send_message.assert_called_once()
@@ -185,7 +185,7 @@ class TestRouterLambdaFunctions(unittest.TestCase):
         )
 
     @mock_sqs
-    @patch("router_lambda_function.os.getenv")
+    @patch("utils_for_filenameprocessor.os.getenv")
     def test_send_to_supplier_queue_success(self, mock_getenv):
         """
         Test send_to_supplier_queue function for a successful message send.
@@ -214,7 +214,7 @@ class TestRouterLambdaFunctions(unittest.TestCase):
         }
 
         # Call the send_to_supplier_queue function
-        success = send_to_supplier_queue(supplier, message_body)
+        success = send_to_supplier_queue(supplier, message_body, sqs)
 
         self.assertTrue(success)
 
@@ -230,7 +230,7 @@ class TestRouterLambdaFunctions(unittest.TestCase):
         )
 
     @mock_sqs
-    @patch("router_lambda_function.os.getenv")
+    @patch("utils_for_filenameprocessor.os.getenv")
     def test_send_to_supplier_queue_queue_not_exist(self, mock_getenv):
         """
         Test send_to_supplier_queue function when the queue does not exist.
@@ -242,6 +242,9 @@ class TestRouterLambdaFunctions(unittest.TestCase):
             "PROD_ACCOUNT_ID": "3456789109",
         }.get(key, default)
 
+        # Create a mock SQS queue
+        mock_sqs = boto3.client("sqs", region_name="eu-west-2")
+
         supplier = "STAR"
         message_body = {
             "vaccine_type": "Flu",
@@ -250,10 +253,12 @@ class TestRouterLambdaFunctions(unittest.TestCase):
             "filename": "Flu_Vaccinations_v5_YGM41_20240708T12130100.csv",
         }
 
-        success = send_to_supplier_queue(supplier, message_body)
+        # Call the send_to_supplier_queue function
+        success = send_to_supplier_queue(supplier, message_body, mock_sqs)
+
         self.assertFalse(success)
 
-    @patch("router_lambda_function.get_supplier_permissions")
+    @patch("initial_file_validation.get_supplier_permissions")
     def test_validate_permissions(self, mock_get_permissions):
         """Test validate vaccine type permissions"""
         mock_get_permissions.return_value = ["FLU_DELETE", "FLU_CREATE", "COVID19_FULL"]
@@ -279,14 +284,14 @@ class TestValidateActionFlagPermissions(unittest.TestCase):
 
         csv_content_dict_reader = convert_string_to_dict_reader(Constants.file_content_with_new_and_delete_action_flags)
 
-        with patch("router_lambda_function.get_supplier_permissions", return_value=allowed_permissions):
+        with patch("initial_file_validation.get_supplier_permissions", return_value=allowed_permissions):
             # Call the function
             result = validate_action_flag_permissions(csv_content_dict_reader, supplier, vaccine_type)
 
         # Check the result
         self.assertTrue(result)
 
-    @patch("router_lambda_function.get_supplier_permissions")
+    @patch("initial_file_validation.get_supplier_permissions")
     def test_validate_action_flag_permissions_with_one_permissions(self, mock_get_supplier_permissions):
         # Mock get_supplier_permissions
         mock_get_supplier_permissions.return_value = ["FLU_DELETE"]
@@ -303,7 +308,7 @@ class TestValidateActionFlagPermissions(unittest.TestCase):
         # Check the result
         self.assertTrue(result)
 
-    @patch("router_lambda_function.get_supplier_permissions")
+    @patch("initial_file_validation.get_supplier_permissions")
     def test_validate_action_flag_permissions_with_full_permissions(self, mock_get_supplier_permissions):
 
         # Mock get_supplier_permissions
@@ -321,7 +326,7 @@ class TestValidateActionFlagPermissions(unittest.TestCase):
         # Check the result
         self.assertTrue(result)
 
-    @patch("router_lambda_function.get_supplier_permissions")
+    @patch("initial_file_validation.get_supplier_permissions")
     def test_validate_action_flag_permissions_with_no_permissions(self, mock_get_supplier_permissions):
 
         # Mock get_supplier_permissions
